requirements.txt
aiohttp==3.10.5
asyncpg==0.29.0
certifi==2024.7.4
click==8.1.7
databases==0.9.0
discord.py==2.4.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-filter==2.0.0
fastapi-pagination==0.12.26
Flask==3.0.3
greenlet==3.0.3
httpx==0.27.0
Hypercorn==0.17.3
Jinja2==3.1.4
mctools==1.3.0
openai==1.35.12
orjson==3.10.6
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
ruamel.yaml==0.18.6
SQLAlchemy==2.0.31
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.4
mysql-connector-python==9.1.0
google-generativeai==0.8.4
google-api-python-client==2.160.0
CurrencyConverter==0.18.2
emoji==2.14.1
cryptography==45.0.7
